requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
pymupdf==1.24.3
python-docx==1.1.2
google-generativeai==0.5.4
python-dotenv==1.0.1
